pyproject.toml
[project]
name = "poly-maker"
version = "0.1.0"
description = "A market making bot for Polymarket prediction markets"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "py-clob-client==0.28.0",
    "python-dotenv==1.2.1",
    "pandas==2.3.3",
    "gspread==6.2.1",
    "gspread-dataframe==4.0.0",
    "sortedcontainers==2.4.0",
    "eth-account==0.13.7",
    "eth-utils==5.3.1",
    "poly_eip712_structs==0.0.1",
    "py_order_utils==0.3.2",
    "requests[socks]==2.32.5",
    "websockets==15.0.1",
    "cryptography==46.0.3",
    "google-auth==2.42.1",
    "web3==7.14.0",
    "InquirerPy==0.3.4",
    "PyYAML==6.0.1",
    "pysocks>=1.7.1",
    "google-generativeai>=0.8.5",
    "feedparser>=6.0.10",
]

[project.optional-dependencies]
dev = [
    "black==24.4.2",
    "pytest==8.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["poly_data", "poly_stats", "poly_utils", "data_updater", "near_sure", "neg_risk_arb", "spike_momentum"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]
